エディタ保存バグ修正、保存内容表示実装
