+++ resources/js/Plugins/styles.css
@@ -46,7 +46,7 @@ h1 {
 }
 
 .editor-container {
-  margin: 20px auto 20px auto;
+  margin: 0px auto 0px auto;
   border-radius: 2px;
   max-width: auto;
   color: #000;
@@ -64,7 +64,6 @@ h1 {
 }
 
 .editor-input {
-  min-height: 150px;
   resize: none;
   font-size: 15px;
   caret-color: rgb(5, 5, 5);
@@ -86,6 +85,17 @@ h1 {
   user-select: none;
   display: inline-block;
   pointer-events: none;
+}
+
+
+.editor-text-superscript {
+  font-size: 0.8em;
+  vertical-align: super;
+}
+
+.editor-text-subscript {
+  font-size: 0.8em;
+  vertical-align: sub
 }
 
 .editor-text-bold {
@@ -355,7 +365,7 @@ pre::-webkit-scrollbar-thumb {
 
 
 .toolbar {
-  display: flex;
+  /*display: flex;*/
   margin-bottom: 1px;
   background: #fff;
   padding: 4px;
@@ -363,6 +373,10 @@ pre::-webkit-scrollbar-thumb {
   border-top-right-radius: 10px;
   vertical-align: middle;
 }
+
+/*.history{*/
+/*  vertical-align: baseline;*/
+/*}*/
 
 .toolbar button.toolbar-item {
   border: 0;
@@ -379,7 +393,7 @@ pre::-webkit-scrollbar-thumb {
 }
 
 .toolbar button.toolbar-item.spaced {
-  margin-right: 2px;
+  margin-right: 4px;
 }
 
 .toolbar button.toolbar-item i.format {
@@ -567,90 +581,5 @@ i.chevron-down {
   margin-right: 12px;
   line-height: 16px;
   background-size: contain;
-}
-
-.link-editor {
-  position: absolute;
-  z-index: 100;
-  top: -10000px;
-  left: -10000px;
-  margin-top: -6px;
-  max-width: 300px;
-  width: 100%;
-  opacity: 0;
-  background-color: #fff;
-  box-shadow: 0px 5px 10px rgba(0, 0, 0, 0.3);
-  border-radius: 8px;
-  transition: opacity 0.5s;
-}
-
-.link-editor .link-input {
-  display: block;
-  width: calc(100% - 24px);
-  box-sizing: border-box;
-  margin: 8px 12px;
-  padding: 8px 12px;
-  border-radius: 15px;
-  background-color: #eee;
-  font-size: 15px;
-  color: rgb(5, 5, 5);
-  border: 0;
-  outline: 0;
-  position: relative;
-  font-family: inherit;
-}
-
-/*.link-editor div.link-edit {*/
-/*  background-image: url(images/icons/pencil-fill.svg);*/
-/*  background-size: 16px;*/
-/*  background-position: center;*/
-/*  background-repeat: no-repeat;*/
-/*  width: 35px;*/
-/*  vertical-align: -0.25em;*/
-/*  position: absolute;*/
-/*  right: 0;*/
-/*  top: 0;*/
-/*  bottom: 0;*/
-/*  cursor: pointer;*/
-/*}*/
-
-.link-editor .link-input a {
-  color: rgb(33, 111, 219);
-  text-decoration: none;
-  display: block;
-  white-space: nowrap;
-  overflow: hidden;
-  margin-right: 30px;
-  text-overflow: ellipsis;
-}
-
-.link-editor .link-input a:hover {
-  text-decoration: underline;
-}
-
-.link-editor .button {
-  width: 20px;
-  height: 20px;
-  display: inline-block;
-  padding: 6px;
-  border-radius: 8px;
-  cursor: pointer;
-  margin: 0 2px;
-}
-
-.link-editor .button.hovered {
-  width: 20px;
-  height: 20px;
-  display: inline-block;
-  background-color: #eee;
-}
-
-.link-editor .button i,
-.actions i {
-  background-size: contain;
-  display: inline-block;
-  height: 20px;
-  width: 20px;
-  vertical-align: -0.25em;
 }
